:sparkles: Change to Service Layer Architecture
